include elf and port arguments
Made a clean-up on the main file.
No modules are loaded and according to user selection,
the respective module is loaded and executed.

Arguments -e --elf and -p --port have being included.
Additionally, the utils module was moved from examples
to the main pyX2Cscope module. So the elf file and port
configuration over config.ini can be called.

A small logic for elf file and port was also added,
see _args_check()