Fix struct enum (#92)
* merged process_end_die
* enum working with structs and arrays

general variable instantiation is now optimized.
Before creating VarInfo objects, a valid address is checked.

This commit also identifies unions. This is not yet concluded because now we have not only byte size, but also bit size and bit offset for union variables smaller than a byte. How to write and read this type of variables is not yet implemented.

* tests pass ok

* enum working and tests ok

* including tests for enum

tests for 16 and 32 bit cores

* optimize offset calculation on struct members

* include test to check struct member address

* union is partially available

if a union is multiple of bytes it is working.
if a union contain elements that have size of some bits, the variable is detected and included, but read and write on these variables are unpredictable. Do not use.

* ruff fix

* version 0.4.2

---------

Co-authored-by: Edras Pacola <edras.pacola@microchip.com>